requirements_dev.txt
pyqt6-tools
